requirements.txt
datasets==2.18.0
django_haystack==3.2.1
pandas==2.2.2
python-dotenv==1.0.1
streamlit==1.32.2
